requirements.txt
streamlit==1.56.0
openai-whisper==20250625
torch==2.5.1
torchaudio==2.5.1
numpy==2.2.6
pydub==0.25.1
soundfile==0.13.1
ffmpeg-python==0.2.0
requests==2.33.1
audio-recorder-streamlit==0.0.10